+++ CHANGELOG.md
@@ -11,6 +11,8 @@ app.
 - Made the puck persist its position, orientation, and active state between
   pages
   ([#1561](https://github.com/birchill/10ten-ja-reader/issues/1561)).
+- Fixed a case of 10ten not looking up text with ruby correctly in Ttsu reader
+  ([#1313](https://github.com/birchill/10ten-ja-reader/issues/1313)).
 - Stopped popup being inverted when Wikipedia's experimental dark mode is
   enabled
   ([#1590](https://github.com/birchill/10ten-ja-reader/issues/1590)).